Dockerfile

+++ main.py
@@ -3,30 +3,18 @@ import os
 from dotenv import load_dotenv
 from pymongo import MongoClient
 from bson.objectid import ObjectId
-from bson.json_util import dumps
+
 
 app = Flask(__name__)
 
 # Load the environ variables
 load_dotenv()
-
-# user_data: dict = {
-#     "name": "Alex",
-#     "username": "Alex@gmail.com",
-#     "password": "Alex!@#1234",
-# }
-#
-# user_data1: dict = {
-#     "name": "Alex Homorazu",
-#     "username": "Alexhom@gmail.com",
-#     "password": "Alex!@#1234",
-# }
-
-
 class DatabaseConnection:
     __conn_string: str | None = os.getenv("MONGO_URI")
 
-    def __init__(self):
+    def __init__(self) -> None:
+        self.connected = False
+
         try:
             # connect the database
             self.client = MongoClient(self.__conn_string)
@@ -36,19 +24,38 @@ class DatabaseConnection:
 
             # get the collection
             self.users_collection = self.db["users"]
+
+            self.connected = True
+            print("Database is connected")
+
         except Exception as e:
-            return f"Mongodb connection error: {e}"
-   
+            print(f"Mongodb connection error: {e}")
+            self.connected = False
+
+
+    def is_connected(self):
+        return self.connected
 
     def initialize_collection(self):
+        if not self.connected:
+            raise Exception("Database not connected")
         self.users_collection.create_index("email", unique=True)
 
+
+    def find_users(self):
+        if not self.connected:
+            raise Exception("Database not connected")
+        return self.users_collection.find({})
     
     def clear_collection(self):
+        if not self.connected:
+            raise Exception("Database not connected")
         self.users_collection.delete_many({})
 
 
     def store_payload(self, payload: dict) -> str:
+        if not self.connected:
+            raise Exception("Database not connected")
         try:
             result = self.users_collection.insert_one(payload)
             print(result)
@@ -58,7 +65,8 @@ class DatabaseConnection:
     
 
     def close_connection(self):
-        self.client.close()
+        if self.client:
+            self.client.close()
 
 
 db = DatabaseConnection()
@@ -69,8 +77,11 @@ db = DatabaseConnection()
 
 @app.route("/")
 def test():
+    if not db.is_connected():
+        return jsonify({
+            "Message": "Database is not connected",
+        }), 500
     return "<p>testing environment</p>"
-
 
 
 @app.get("/users")
@@ -78,8 +89,12 @@ def fetch_users():
     """
     fetch all the users from the database
     """
-    # if request.method == "GET":
-    users = db.users_collection.find({})
+    if not db.is_connected():
+        return jsonify({
+            "Message": "Database is not connected",
+        }), 500
+
+    users = db.find_users()
     results = []
     for user in users:
         user["_id"] = str(user["_id"])
@@ -89,6 +104,11 @@ def fetch_users():
 
 @app.post("/users")
 def post_user():
+    if not db.is_connected():
+        return jsonify({
+            "Message": "Database is not connected",
+        }), 500
+
     req_data = request.get_json()
     name = req_data["name"]
     username = req_data["username"]
@@ -109,7 +129,13 @@ def get_user(id: str):
     """
     fetch a specific user from the database provide user_id
     """
+
     try:
+        if not db.is_connected():
+            return jsonify({
+                "Message": "Database is not connected",
+            }), 500
+
         user = db.users_collection.find_one({"_id": ObjectId(id)})
         if user:
             user["_id"] = str(user["_id"])
@@ -124,27 +150,46 @@ def get_user(id: str):
 
 @app.delete("/users/<id>")
 def delete_user(id: str):
-    db.users_collection.delete_one({"_id": ObjectId(id)})
-    return "Deleted user successfully"
+    try:
+        if not db.is_connected():
+            return jsonify({
+                "Message": "Database is not connected",
+            }), 500
+
+        db.users_collection.delete_one({"_id": ObjectId(id)})
+        return jsonify({"Message": "Deleted user successfully"})
+    except Exception as e:
+        return jsonify({
+            "Message": str(e)
+        }), 500
 
 @app.put("/users/<id>")
 def update_user(id: str):
-    req = request.get_json()
-    query_filter = {
-        "_id": ObjectId(id), 
-    }
+    try:
+        if not db.is_connected():
+            return jsonify({
+                "Message": "Database is not connected",
+            }), 500
 
-    update_operation = {
-        "$set": {
-            "name": req["name"],
-            "username": req["username"],
-            "password": req["password"],
+        req = request.get_json()
+        query_filter = {
+            "_id": ObjectId(id), 
         }
-    }
-    db.users_collection.update_one(query_filter, update_operation)
-    return f"Update successful"
 
+        update_operation = {
+            "$set": {
+                "name": req["name"],
+                "username": req["username"],
+                "password": req["password"],
+            }
+        }
+        db.users_collection.update_one(query_filter, update_operation)
+        return jsonify({"Message":"Update successful"})
+    except Exception as e:
+        return jsonify({
+            "Message": str(e),
+        }), 500
 # db.close_connection()
 
 if __name__ == "__main__":
-    app.run(debug=True)
+    app.run(host="0.0.0.0", port=3000, debug=True)
